fix : task 3 (bug : repeated values are printing) - resolved feat : task 1 : added query to print the lis tin descending order
